+++ src/index.js
@@ -6,7 +6,7 @@ import './js/modalClose';
 import './js/firstPageLoad.js';
 import './js/pagination';
 import './js/modal_our_team';
-import './js/renderModal.js';
+import './js/modalOpen.js';
 import './js/morefromauthor.js';
 import './js/up_button.js';
