+++ src/js/morefromauthor.js
@@ -3,7 +3,7 @@ import onCloseMainModal from './modalClose.js';
 import eventServiceApi from './search-API';
 import templateCard from '../templates/cardModif.hbs';
 import { setPagination } from './pagination.js';
-import { authorName } from './renderModal.js';
+import { authorName } from './modalOpen.js';
 
 refs.buttonMoreFromAuthor.addEventListener('click', searchMoreFromAuthor);
